rename: 패키직 구조 변경
